自定义Security登录逻辑 合并 创建实例和保存实例内容接口 修改第三方登陆逻辑 使用uuid创建实例id

+++ springboot-dubbo-api/src/main/java/com/lzq/api/dto/ExampleAccount.java
@@ -53,7 +53,7 @@ public class ExampleAccount implements Serializable {
      */
     @TableId(value = "example_id")
     @JsonProperty("exampleId")
-    private Integer exampleId;
+    private Long exampleId;
     /**
      * 实列名称
      */
